Make boot() very fast on resume.
We always start the VM with a frozen CPU. For normal boots, this just
starts the qemu process during NewVM, but only unpauses the CPU during
Start. The only net effect is that we allocate system memory during
NewVM, which is actually a net improvement.

For resumes however, this makes a big difference: we start the VM
during thaw and make it load its snapshot, which takes several
seconds. We can parallelize that loading. Then, once everything is
loaded, we can blow through everything and unpause in one big batch,
so that the clock skew on the VMs is minimized.