Parallelize VM thaw, wait for all thawed before unpausing.
This should make the clock skew minimal, because we let all VMs work
to reload themselves into memory, but wait for all of them to reach
the "loaded, but paused" state. Then we quickly blow through all of
them and unpause their CPUs, so they all restart execution at roughly
the same time.