Actively enforce VM system time on boot and resume.
Because all the clocks are freewheeling now, if we boot
two VMs 5 minutes apart, the most recently booted VM will
be 5 minutes behind the older VM... Which makes TLS certs
very sad.

So, we now explicitly track how much time has elapsed in
the simulated universe (accounting for suspends and resumes),
and whenever a VM first boots or resumes from sleep, we actively
force it to be at the time of the universe.

In practice this means all VMs in the universe have slightly
different times because there's no external clock discipline
driving them all together, but the difference is going to be
sub-second, which is good enough for our uses.